Subject: Update ignored files list and remove explicit dependencies from README
This commit updates the `.gitignore` file to include more IDE/system specific files such as `.venv`, `.mypy_cache`, `.idea`, and `.DS_Store`. This helps to prevent these system and environment specific files from being accidentally committed to the repository.

In `README.md`, the explicit list of dependencies has been replaced with a reference to a `requirements.txt` file. This simplifies the update process of dependencies and ensures only one authoritative source for them.

There is also a minor change involving the replacement of similar code lines for consistency and readability in `main.py`.

All changes aim to improve the general maintenance and readability of the project.